// dashboard-v6/src/components/api/Corpus.ts
export interface ISimSent {
  sent: string;
  sim: number;
}
export interface ISentenceSimListResponse {
  ok: boolean;
  message: string;
  data: { rows: ISimSent[]; count: number };
}